Extract dataloader utilities from `TrainerDataLoadingMixin` (#10145)
